backend/requirements.txt
fastapi==0.122.0
langchain_core==1.1.0
langchain_google_genai==3.2.0
pydantic==2.12.5
python-dotenv==1.2.1
uvicorn==0.38.0
